requirements.txt
flask
pandas
numpy
PyYAML
nltk
scikit-learn